test: modify the test for allowing injection of custom HttpMessageHandler
